Refactor nearest() for better efficiency
The effectiveness of the pruning rule depends on quickly finding a nearby point.
To do this, organize your recursive method so that when there are two possible subtrees to go down, you always choose the subtree that is on the same side of the splitting line as the query point as the first subtree to explore—the closest point found while exploring the first subtree may enable pruning of the second subtree.